+++ selfdrive/debug/profiling/perfetto/server.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/bash
+
+curl -LO https://get.perfetto.dev/trace_processor
+chmod +x ./trace_processor
+
+./trace_processor --httpd