+++ selfdrive/ui/qt/widgets/ssh_keys.cc
@@ -41,23 +41,22 @@ void SshControl::refresh() {
 
 void SshControl::getUserKeys(const QString &username) {
   HttpRequest *request = new HttpRequest(this, false);
-  QObject::connect(request, &HttpRequest::receivedResponse, [=](const QString &resp) {
-    if (!resp.isEmpty()) {
-      params.put("GithubUsername", username.toStdString());
-      params.put("GithubSshKeys", resp.toStdString());
+  QObject::connect(request, &HttpRequest::requestDone, [=](const QString &resp, bool success) {
+    if (success) {
+      if (!resp.isEmpty()) {
+        params.put("GithubUsername", username.toStdString());
+        params.put("GithubSshKeys", resp.toStdString());
+      } else {
+        ConfirmationDialog::alert(QString("Username '%1' has no keys on GitHub").arg(username), this);
+      }
     } else {
-      ConfirmationDialog::alert("Username '" + username + "' has no keys on GitHub", this);
+      if (request->timeout()) {
+        ConfirmationDialog::alert("Request timed out", this);
+      } else {
+        ConfirmationDialog::alert(QString("Username '%1' doesn't exist on GitHub").arg(username), this);
+      }
     }
-    refresh();
-    request->deleteLater();
-  });
-  QObject::connect(request, &HttpRequest::failedResponse, [=] {
-    ConfirmationDialog::alert("Username '" + username + "' doesn't exist on GitHub", this);
-    refresh();
-    request->deleteLater();
-  });
-  QObject::connect(request, &HttpRequest::timeoutResponse, [=] {
-    ConfirmationDialog::alert("Request timed out", this);
+
     refresh();
     request->deleteLater();
   });